增加使用setuptools编译例子
1.增加使用setuptools编译例子
2.setuptools 已经基本完成可以编译 动态 静态 以及 可执行文件
3.修复getrealpath 在 macos下会获取软连接地址的错误
4.增加isset 判断是否存在变量,可以再dofile 跨文件中来判断并且不出warning警告.

+++ lib/socket/stdafx.h
@@ -16,12 +16,12 @@ inline std::string cpps_socket_g2u(std::string gb2312)
 {
 	std::string ret = "";
 	int len = MultiByteToWideChar(CP_ACP, 0, gb2312.c_str(), -1, NULL, 0);
-	wchar_t* wstr = new wchar_t[len + 1];
-	memset(wstr, 0, len + 1);
+	wchar_t* wstr = new wchar_t[(size_t)len + 1];
+	memset(wstr, 0, (size_t)len + 1);
 	MultiByteToWideChar(CP_ACP, 0, gb2312.c_str(), -1, wstr, len);
 	len = WideCharToMultiByte(CP_UTF8, 0, wstr, -1, NULL, 0, NULL, NULL);
-	char* str = (char*)malloc(len + 1);
-	memset(str, 0, len + 1);
+	char* str = (char*)malloc((size_t)len + 1);
+	memset(str, 0, (size_t)len + 1);
 	WideCharToMultiByte(CP_UTF8, 0, wstr, -1, str, len, NULL, NULL);
 	ret = str;
 	if (wstr) delete[] wstr;
